+++ MidnightSimpleUnitFrames/Libs/MSUFUnitFrames/MSUF_UF_Core.lua
@@ -118,7 +118,6 @@ local STATUS_EVENT_ELEMENTS = {
   IncomingResIndicator = true,
   PVPIndicator = true,
   StanceIndicator = true,
-  TargetingYouIndicator = true,
 }
 
 local IDENTITY_ELEMENTS = {
@@ -142,7 +141,6 @@ local IDENTITY_ELEMENTS = {
   CombatIndicator = true,
   IncomingResIndicator = true,
   PVPIndicator = true,
-  TargetingYouIndicator = true,
 }
 UF.identityElements = IDENTITY_ELEMENTS