+++ MidnightSimpleUnitFrames/State/MSUF_Profiles.lua
@@ -2143,7 +2143,6 @@ local MSUF_PROFILEIO_STATUS_PREFIXES = {
     "incomingResIndicator",
     "pvpIndicator",
     "stanceIndicator",
-    "targetingYouIndicator",
     "raidGroupName",
 }
 local MSUF_PROFILEIO_GROUP_STATUS_NUMERIC_KEYS = {